pyproject.toml
[tool.poetry]
name = "keepem"
version = "0.1.0"
description = "Keepem backends "
authors = ["Marti <marti@keepem.app>"]
license = "Coppyright Keepem"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.70.1"
uvicorn = "0.15.0"
pymongo = "4.0.1"
fastapi-pagination = "0.9.1"
python-dotenv = "0.19.2"
pydantic = {version = "1.9.0", extras = ["email", "dotenv"]}
python-multipart = "0.0.5"
aiofiles = "0.8.0"
jinja2 = "3.0.3"
bcrypt = "3.2.0"
flatdict = "4.0.1"
gitpython = "3.1.24"
pyinstrument = "4.1.1"
setuptools-rust = "1.1.2"
passlib = "1.7.4"
boto3 = "1.20.24"

cryptography = "3.4.7"
python-jose = {version = '3.3.0', extras = ["cryptography"]}
pyCryptodome = "3.10.1"
pyAesCrypt = "6.0.0"

[tool.poetry.dev-dependencies]
pytest = "6.2.4"
pytest-xdist = "2.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
